requirements.txt
newspaper3k
nltk
gtts
lxml==4.9.3



# pip install -r requirements.txt
